requirements.txt
Django==1.8.2
argparse==1.2.1
djangorestframework==3.1.2
wsgiref==0.1.2
